Bug fixes and keypress functionality
Fixed a couple of bugs, still a couple of small ones outstanding. I've also added keyboard functionality for the numbers and the decimal button. Happy to revisit this and add functionality for the other buttons at a later date, but at this stage I think my time is best served moving on.